% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_activity.R
\name{activity_empirical}
\alias{activity_empirical}
\title{Retrieve fitted activity scores for a given setting}
\usage{
activity_empirical(setting_string, n_classes)
}
\arguments{
\item{setting_string}{A string indicating the setting}

\item{n_classes}{The number of activity classes}
}
\value{
A numeric vector of activity scores
}
\description{
This function retrieves the fitted activity scores for a specific setting
based on empirical data from (Pung et al., 2024) Interface
}
\examples{
activity_empirical("haslemere", n_classes = 5)
}
