% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_models.R
\name{activity_linear}
\alias{activity_linear}
\title{Generate the activity-class scaler for each activity class
Linear activity scores}
\usage{
activity_linear(min_act, max_act, n_classes)
}
\description{
Simple function which linearly interpolates between the lowest-activity and
the highest-activity class. Hence, for each class i we have a positive number
A_i which gives a measure of activity.
}
\details{
(More complex models of how activity
changes with the activity class are possible.)
}
