% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_models.R
\name{activity_null}
\alias{activity_null}
\title{Generate a null activity vector}
\usage{
activity_null(n_classes)
}
\arguments{
\item{n_classes}{Number of activity classes}
}
\value{
A numeric vector of activity scores (all equal to 1)
}
\description{
Function of use when no activity structure is desired.
Especially relevant for testing purposes.
}
