% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_models.R
\name{adjacent_switching}
\alias{adjacent_switching}
\title{Switching matrix where you can only move between adjacent classes}
\usage{
adjacent_switching(switch_rate, num_classes)
}
\arguments{
\item{switch_rate}{switcjing rate. in general = 1/expected residence time
for the non-extreme classes (ie, 2,3,4) for a 5-class model.}

\item{num_classes}{number of activity classes}
}
\description{
Hence there is twice expected residence time in the extreme classes.
}
\seealso{
\code{\link{simple_switching}} which uses uniform movements
where you can move more than distance of 1 class at once.
}
