% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_models.R
\name{adjacent_switching_wrapped}
\alias{adjacent_switching_wrapped}
\title{Switching matrix where you can only move between adjacent classes but wraps lowest and highest activities.}
\usage{
adjacent_switching_wrapped(switch_rate, num_classes)
}
\arguments{
\item{switch_rate}{switching rate. in general = 1/expected residence time
for the non-extreme classes (ie, 2,3,4) for a 5-class model.}

\item{num_classes}{number of activity classes}
}
\description{
Note that we can have min activity -> max activity in a single step.
}
