% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximations.R
\name{approx_both}
\alias{approx_both}
\title{Approximate a switching system with a no-switch system
using both minimally constrained contact matrix and
epsilon-constrained approximations.}
\usage{
approx_both()
}
\description{
Approximate a switching system with a no-switch system
using both minimally constrained contact matrix and
epsilon-constrained approximations.
}
