% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximations.R
\name{approx_epsilon}
\alias{approx_epsilon}
\title{Apperoximate a switching system with a no-switch system
using a single scalar epsilon parameter for assortativity/proportionate
mixing.}
\usage{
approx_epsilon(reference_parameters, initial_condition_rule, pathogen_string)
}
\arguments{
\item{reference_parameters}{A list of reference parameters.}
}
\description{
Apperoximate a switching system with a no-switch system
using a single scalar epsilon parameter for assortativity/proportionate
mixing.
}
