% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximations.R
\name{approx_minimally_constrained}
\alias{approx_minimally_constrained}
\title{Approximate a switching system with a minimally constrained contact matrix
with no switching}
\usage{
approx_minimally_constrained()
}
\description{
Approximate a switching system with a minimally constrained contact matrix
with no switching
}
\seealso{
approx_epsilon which approximates
with a single scalar epsilon parameterising
the distribution.
}
