% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximations.R
\name{approximation_error}
\alias{approximation_error}
\title{Characterise error in approximations across
parameter space}
\usage{
approximation_error(gamma_range, switch_rate_range)
}
\arguments{
\item{gamma_range}{range of recovery rates to explore}

\item{switch_rate_range}{range of switching rates to explore}
}
\description{
Characterise error in approximations across
parameter space
}
