% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_curves.R
\name{bar_comparison}
\alias{bar_comparison}
\title{Bar comparison of outcome metrics across models}
\usage{
bar_comparison(comparison_df, palette)
}
\arguments{
\item{comparison_df}{the dataframe of comparison metrics generated by make_metric_comparison,
which should include columns for model_type, ptEI, psEI, ptI, psI, hit, final_size, final_time, and r0.
The function pivots this dataframe into a long format suitable for ggplot2,
where each row corresponds to a specific metric for a given model type.}

\item{palette}{gives the colour palette to use for each model type.
Should be a vector of colours of length equal to the number of model
types (rows) in comparison_df. Can be generated with make_model_palette.}
}
\description{
Bar comparison of outcome metrics across models
}
