% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_store.R
\name{basic_scalar_parms}
\alias{basic_scalar_parms}
\title{Get the scalar epi parameters for given pathogen}
\usage{
basic_scalar_parms(pathogen_string = "sars-cov-2")
}
\arguments{
\item{pathogen_string}{string describing the desired pathogen
for which the parameters
are required. Default value is "sars-cov-2" for which we use the parameterisation
from (Britton et al., 2020) Science}
}
\value{
a list of scalars whose elements are \verb{$beta}, \verb{$sigma}, \verb{$gamma}
}
\description{
Convenience function which returns the scalar parameters
(ie transmission rate (beta),
rate of progression to I (sigma) and the recovery rate (gamma).
}
\concept{parameterisation}
