% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_store.R
\name{calibrate_parms}
\alias{calibrate_parms}
\title{Changes the β value until the target R0 is met}
\usage{
calibrate_parms(preliminary_parms, target_r0)
}
\arguments{
\item{preliminary_parms}{list of preliminary parameters
including mixing matrix,
activity scores, class sizes, etc.}

\item{target_r0}{desired R0 value}
}
\value{
updated parameters with β set to achieve target R0
}
\description{
Uses a very simple approach by dividing the target R0 by the current R0
and scaling β by that factor.
}
\concept{calibration}
