% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_metrics.R
\name{characterise_end}
\alias{characterise_end}
\title{Compute epidemic duration and final size}
\usage{
characterise_end(model_output, classed_detail = TRUE)
}
\arguments{
\item{model_output}{Model output}

\item{classed_detail}{Whether to compute per-class results}
}
\description{
Compute epidemic duration and final size
}
\seealso{
Other epidemic_metrics: 
\code{\link{characterise_peaks}()},
\code{\link{compute_all}()},
\code{\link{compute_hit}()},
\code{\link{compute_rt}()},
\code{\link{compute_spectral_gap}()},
\code{\link{compute_ttprev}()},
\code{\link{find_peaks}()},
\code{\link{scale_results}()}
}
\concept{epidemic_metrics}
