% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_metrics.R
\name{characterise_peaks}
\alias{characterise_peaks}
\title{Characterise epidemic peaks for total and by class}
\usage{
characterise_peaks(model_output, info, classed_detail = TRUE, pars)
}
\arguments{
\item{model_output}{Model output (scaled)}

\item{info}{Info list from get_info()}

\item{classed_detail}{If TRUE, compute for every activity class}
}
\description{
This takes in scaled epidemic output and computes peak sizes.
}
\seealso{
Other epidemic_metrics: 
\code{\link{characterise_end}()},
\code{\link{compute_all}()},
\code{\link{compute_hit}()},
\code{\link{compute_rt}()},
\code{\link{compute_spectral_gap}()},
\code{\link{compute_ttprev}()},
\code{\link{find_peaks}()},
\code{\link{scale_results}()}
}
\concept{epidemic_metrics}
