% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vary_parameters.R
\name{check_fraction_exposed}
\alias{check_fraction_exposed}
\title{Sanitise the exposed fraction}
\usage{
check_fraction_exposed(fraction_exposed)
}
\arguments{
\item{fraction_exposed}{the fraction of the pop
which is put in the E compartment(s)}
}
\description{
Makes sure that the exposed fraction is between
0 and 1.
}
