% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_onto_noswitch.R
\name{compare_approx}
\alias{compare_approx}
\title{Function to plot the approximated system against the original system}
\usage{
compare_approx(ref_parms, approx_parms, model)
}
\arguments{
\item{shared_parms}{the parameters which are shared by both the approxi-
mate system and the exact system. Usually this is just the  scalar parameters}

\item{original_parms}{the original parameters: usually nonzero switching matrix
and the contact matrix.}

\item{approximate_parms}{the optimised parameters giving similar epidemic
trajectories as the original parameters}
}
\description{
Allows for immediate visual comparison of the trajectory of the
possible epidemic, both with and without movemement between activity
classes. This function plots the trajectories next to each other.
}
\details{
It also computes the
}
