% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_metrics.R
\name{compute_all}
\alias{compute_all}
\title{Wrapper function to compute all metrics}
\usage{
compute_all(
  model_output,
  classed_detail = TRUE,
  prevalence_threshold = NULL,
  params
)
}
\arguments{
\item{model_output}{Model output (must contain Rt column or
params must be provided)}

\item{prevalence_threshold}{Prevalence threshold for ttprev}

\item{params}{Model parameters used to generate model output}
}
\description{
Wrapper function to compute all metrics
}
\seealso{
Other epidemic_metrics: 
\code{\link{characterise_end}()},
\code{\link{characterise_peaks}()},
\code{\link{compute_hit}()},
\code{\link{compute_rt}()},
\code{\link{compute_spectral_gap}()},
\code{\link{compute_ttprev}()},
\code{\link{find_peaks}()},
\code{\link{scale_results}()}
}
\concept{epidemic_metrics}
