% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_metrics.R
\name{compute_hit}
\alias{compute_hit}
\title{Compute empirical herd immunity threshold (HIT)}
\usage{
compute_hit(model_output_rt)
}
\arguments{
\item{model_output_rt}{Model output data.table
with Rt column already calculated}
}
\value{
A list with elements:
\itemize{
\item HIT: the empirical herd immunity threshold
\item HIT_time: the time when Rt crosses below 1
\item Rt_at_HIT: the value of Rt at the HIT time (should be close to 1)
}
}
\description{
Computes the empirical herd immunity threshold (HIT). This is
defined as the proportion of the population that has been infected when the
effective reproduction number Rt crosses below 1 for the first time.
}
\seealso{
Other epidemic_metrics: 
\code{\link{characterise_end}()},
\code{\link{characterise_peaks}()},
\code{\link{compute_all}()},
\code{\link{compute_rt}()},
\code{\link{compute_spectral_gap}()},
\code{\link{compute_ttprev}()},
\code{\link{find_peaks}()},
\code{\link{scale_results}()}
}
\concept{epidemic_metrics}
