% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_metrics.R
\name{compute_homogeneous_metrics}
\alias{compute_homogeneous_metrics}
\title{Computes the outcome metrics for a homogeneous population}
\usage{
compute_homogeneous_metrics(base_params)
}
\arguments{
\item{base_params}{the basic parameters: beta, sigma, gamma}
}
\description{
That is, the outcome metrics for a population where
each individual has the same activity. This is equivalent
to a mass action single pop.
}
