% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_metrics.R
\name{compute_ngm}
\alias{compute_ngm}
\title{Constructs the next-generation matrix}
\usage{
compute_ngm(pars)
}
\arguments{
\item{pars}{The model parameters (list)}
}
\description{
Make the next-generation matrix (NGM) for the specified model. This
uses the usual next-generation approach for compartmental models with a
large domain.
}
\seealso{
Other r0_calc: 
\code{\link{compute_r0}()},
\code{\link{sigma_matrix}()}
}
\concept{r0_calc}
