% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_metrics.R
\name{compute_r0}
\alias{compute_r0}
\title{Computes the next-generation matrix and R0}
\usage{
compute_r0(pars)
}
\arguments{
\item{pars}{the parameters for the model run}
}
\description{
Computes the next-generation matrix and R0
}
\seealso{
Other r0_calc: 
\code{\link{compute_ngm}()},
\code{\link{sigma_matrix}()}
}
\concept{r0_calc}
