% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_metrics.R
\name{compute_rt}
\alias{compute_rt}
\title{Computes the Rt using the population state at a particular time t}
\usage{
compute_rt(pars, popstate)
}
\arguments{
\item{pars}{the model parameters}

\item{popstate}{the state of the population at the time for Rt.
Must have a list element $s_sizes giving the number of people
who are susceptible to infection in each activity class. This
will then be divided through by the vector of activity class
sizes.}
}
\description{
Computes the Rt using the population state at a particular time t
}
\seealso{
\code{compute_r0}

Other epidemic_metrics: 
\code{\link{characterise_end}()},
\code{\link{characterise_peaks}()},
\code{\link{compute_all}()},
\code{\link{compute_hit}()},
\code{\link{compute_spectral_gap}()},
\code{\link{compute_ttprev}()},
\code{\link{find_peaks}()},
\code{\link{scale_results}()}
}
\concept{epidemic_metrics}
