% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_metrics.R
\name{compute_rt_ts}
\alias{compute_rt_ts}
\title{Compute R0 and Rt time series from model output}
\usage{
compute_rt_ts(model_output, pars, add_to_output = TRUE)
}
\arguments{
\item{model_output}{Model output from dust/odin}

\item{pars}{Model parameters}

\item{add_to_output}{If TRUE, adds R0 and Rt columns to model_output}
}
\value{
If add_to_output is TRUE, returns model_output with R0 and Rt
columns added. Otherwise, returns a data.frame with time, R0, Rt columns.
}
\description{
Uses an effective NGM approach, accounting for the depletion of susceptibles
}
