% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_metrics.R
\name{compute_spectral_gap}
\alias{compute_spectral_gap}
\title{spectral gap calculation}
\usage{
compute_spectral_gap(pars)
}
\arguments{
\item{pars}{parameters for model run}
}
\description{
compute the difference in absolute value between the
dominant and second-most dominant eigenvals.
Should provide some measure of the time of duration
transient behaviour.
}
\seealso{
Other epidemic_metrics: 
\code{\link{characterise_end}()},
\code{\link{characterise_peaks}()},
\code{\link{compute_all}()},
\code{\link{compute_hit}()},
\code{\link{compute_rt}()},
\code{\link{compute_ttprev}()},
\code{\link{find_peaks}()},
\code{\link{scale_results}()}
}
\concept{epidemic_metrics}
