% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_metrics.R
\name{compute_ttprev}
\alias{compute_ttprev}
\title{Compute time to threshold prevalence}
\usage{
compute_ttprev(
  model_output,
  prevalence_threshold,
  num_classes,
  classed_detail = TRUE
)
}
\arguments{
\item{model_output}{Output from dust/odin}

\item{prevalence_threshold}{Numeric threshold for (E + I) / N_j}

\item{num_classes}{Number of activity classes}

\item{classed_detail}{Whether to compute per class}
}
\description{
Compute time to threshold prevalence
}
\seealso{
Other epidemic_metrics: 
\code{\link{characterise_end}()},
\code{\link{characterise_peaks}()},
\code{\link{compute_all}()},
\code{\link{compute_hit}()},
\code{\link{compute_rt}()},
\code{\link{compute_spectral_gap}()},
\code{\link{find_peaks}()},
\code{\link{scale_results}()}
}
\concept{epidemic_metrics}
