% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_curves.R
\name{curve_comparison}
\alias{curve_comparison}
\title{Compare SEIR curves across models, using the palette to colour the curves by model type.
This is a more direct comparison of the epidemic trajectories than the bar_comparison of summary
metrics, and can reveal differences in timing and shape of the epidemics that may not be
captured by summary metrics alone.
Facets by S, e, i ,r and plots the occupancy for each of those states for each model .}
\usage{
curve_comparison(parsets, palette, show_homlines = FALSE)
}
\arguments{
\item{parsets}{the list of parameter sets to compare, which should include the calibrated
parameters and model names for each model type.}

\item{palette}{a vector of colours to use for each model type, of the same length
as the number
of model types in parsets. Can be generated with make_model_palette.}

\item{show_homlines}{a boolean indicating whether to show the homogeneous epidemic lines
in the bar plots}
}
\description{
Compare SEIR curves across models, using the palette to colour the curves by model type.
This is a more direct comparison of the epidemic trajectories than the bar_comparison of summary
metrics, and can reveal differences in timing and shape of the epidemics that may not be
captured by summary metrics alone.
Facets by S, e, i ,r and plots the occupancy for each of those states for each model .
}
