% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{declass}
\alias{declass}
\title{Simplifies epidemic output}
\usage{
declass(classed_epidf)
}
\arguments{
\item{classed_epidf}{epidemic with class structure}
}
\value{
epidemic with S, E, I, R and time columns
}
\description{
Especially useful when comparing epidemic trajectories across
homogeneous-inhomogeneous parameter sets
}
