% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{display_mix_switch}
\alias{display_mix_switch}
\title{Pretty-print contact and switching matrices}
\usage{
display_mix_switch(mix_mat, swtch_mat)
}
\arguments{
\item{mix_mat}{the contact or mixing matrix describing intra-
and inter-activity class contatc frequencies}

\item{swtch_mat}{the matrix describing the switching of hosts between}
}
\description{
Pretty-print the mixing and switching matrix
}
