% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_models.R
\name{elaborated_switching}
\alias{elaborated_switching}
\title{Elaborated switching}
\usage{
elaborated_switching(switch_rate, num_classes, increment_scalar)
}
\description{
Here we assume that w(i,j) is the flow from i -> j. Note this is different
to some other conventions.
Particularly, the ngm calculation using Sigma matrix.
}
