% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_plots.R
\name{example_comp}
\alias{example_comp}
\title{Example comparison scan and plot}
\usage{
example_comp(res = 50, target_r0, pathogen_string)
}
\arguments{
\item{res}{grid resolution}

\item{target_r0}{the target r0 to calibrate against}
}
\description{
Scans
}
