% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_models.R
\name{exponential_mixing}
\alias{exponential_mixing}
\title{Parameterically generate exponential mixing model}
\usage{
exponential_mixing(n_activity, epsilon_assort, act_vector, class_size_vector)
}
\arguments{
\item{n_activity}{the number of activity classes}

\item{epsilon_assort}{the assortativity of the contact pattern}

\item{act_vector}{the vector of activity class scores}

\item{class_size_vector}{the vector of activity class sizes}
}
\description{
Automatically computes the normalisation
constants needed
}
