% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_metrics.R
\name{find_peaks}
\alias{find_peaks}
\title{Find epidemic peaks (for I or E + I)}
\usage{
find_peaks(epi_data, s_label, e_label, i_label, pop_size, class, pars)
}
\arguments{
\item{epi_data}{data.table with epidemic trajectories}

\item{s_label}{Column name for susceptibles}

\item{e_label}{Column name for exposed}

\item{i_label}{Column name for infected}

\item{pop_size}{Total (sub)population size}

\item{class}{Activity class identifier}
}
\description{
This function requires that the data are scaled.
}
\seealso{
Other epidemic_metrics: 
\code{\link{characterise_end}()},
\code{\link{characterise_peaks}()},
\code{\link{compute_all}()},
\code{\link{compute_hit}()},
\code{\link{compute_rt}()},
\code{\link{compute_spectral_gap}()},
\code{\link{compute_ttprev}()},
\code{\link{scale_results}()}
}
\concept{epidemic_metrics}
