% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_models.R, R/utilities.R
\name{find_steadystate}
\alias{find_steadystate}
\title{Find the steady-state distribution of a continuous-time Markov chain
given its transition rate matrix W.}
\usage{
find_steadystate(W)

find_steadystate(W)
}
\arguments{
\item{W}{A square matrix where W(i,j) is the rate of transition}
}
\description{
Find the steady-state distribution of a continuous-time Markov chain
given its transition rate matrix W.

Find the steady-state distribution of a continuous-time Markov chain
given its transition rate matrix W.
}
