% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_onto_noswitch.R
\name{fit_noswitch}
\alias{fit_noswitch}
\title{Fit non-switching system}
\usage{
fit_noswitch(
  parms,
  model,
  eps_init = NULL,
  metric = c("i_tot", "e_tot"),
  normalize = TRUE,
  verbose = TRUE,
  nl_opts = list(algorithm = "NLOPT_LN_BOBYQA", maxeval = 200, xtol_rel = 1e-06)
)
}
\arguments{
\item{parms}{a parameterisation for the modelled system
which includes switching rate > 0 (can also be zero for testing)}

\item{model}{odin model used for simulation}

\item{eps_init}{initial value of epsilon which is used for the reference
system.}

\item{metric}{the way}
}
\value{
sol solution of the optimisation problem
}
\description{
Produce a fitted time-invariant system which is the 'best' approximation
to a time-varying epidemic system.
}
\details{
This function produces, for a given parameterisation of the activity class
switching matrix, the mixing/contact matrix producing the
most similar results for the
Exposed compartment.
}
