% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_helpers.R
\name{generate_mixingmat}
\alias{generate_mixingmat}
\title{Generate a mixing/contact matrix with selectable model}
\usage{
generate_mixingmat(
  n_activity,
  epsilon_assort,
  act_vector,
  class_size_vector,
  mixing_model = c("garnett", "polynomial", "exponential"),
  power_param = NULL
)
}
\arguments{
\item{n_activity}{integer number of activity classes}

\item{epsilon_assort}{assortativity parameter}

\item{act_vector}{vector of activity scores}

\item{class_size_vector}{vector of class sizes}

\item{mixing_model}{character; one of 'garnett', 'polynomial', 'exponential'}
}
\description{
Wrapper to produce a mixing matrix. Defaults to the Garnett model.
}
