% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_metrics.R
\name{get_info}
\alias{get_info}
\title{Convenience function to get population info from model output}
\usage{
get_info(model_output)
}
\arguments{
\item{model_output}{Model output data.table}
}
\value{
info, a list containing relevant information
}
\description{
Gets number of activity classes and total population size.
}
\concept{epidemic_info}
