% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_curves.R
\name{homogeneous_curve}
\alias{homogeneous_curve}
\title{Plot homogeneous mass action model}
\usage{
homogeneous_curve(r0)
}
\arguments{
\item{r0}{the target r0 for the homogeneous model.
Note that this uniquely defines the epidemic given we have
fixed params}

\item{n_activity}{the number of activity classes
involved.}
}
\value{
a ggplot2 object and saves output to pdf file
}
\description{
Plot homogeneous mass action model
}
