% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_curves.R
\name{homogeneous_metrics}
\alias{homogeneous_metrics}
\title{Metrics for homogeneous SEIR epidemic}
\usage{
homogeneous_metrics(r0)
}
\arguments{
\item{r0}{Target basic reproduction number}
}
\value{
A named list with ptI, psI, ptEI, psEI,
hit, final_size, end_time, r0
}
\description{
Computes peak times/sizes, final size, HIT and returns a metrics list
for the homogeneous (mass-action) SEIR epidemic defined by target R0.
}
