% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{kronecker_delta}
\alias{kronecker_delta}
\title{kronecker_delta}
\usage{
kronecker_delta(i, j)
}
\arguments{
\item{i}{first index}

\item{j}{second index}
}
\value{
1 if i == j, 0 otherwise
}
\description{
kronecker_delta
}
