% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_store.R
\name{make_contact_matrix}
\alias{make_contact_matrix}
\title{Construct the contact matrix}
\usage{
make_contact_matrix(
  n_activity,
  epsilon,
  activity_vector,
  class_size_vector,
  power_param = NULL,
  mixing_model
)
}
\arguments{
\item{n_activity}{the number of activity groups in the model}

\item{epsilon}{the assortativity for the contact/mixing model}

\item{activity_vector}{the vector of activity scores}

\item{class_size_vector}{the vector of class sizes in the model}

\item{power_param}{the power parameter for}

\item{mixing_model}{the model of contact/mixing}
}
\description{
Construct the contact matrix
}
\concept{contact_models}
