% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_onto_noswitch.R
\name{make_generator}
\alias{make_generator}
\title{Make a generator matrix}
\usage{
make_generator(prelim_matrix, diag_element)
}
\arguments{
\item{prelim_matrix}{A preliminary matrix.}

\item{diag_element}{The diagonal element value.}
}
\description{
Makes a generator for the switching process.
}
