% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_plots.R
\name{make_homogeneous_parms}
\alias{make_homogeneous_parms}
\title{Make homogeneous parameters}
\usage{
make_homogeneous_parms(base_params)
}
\arguments{
\item{base_params}{base parameters with potential heterogeneity}
}
\description{
Make homogeneous parameters
}
