% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_store.R
\name{make_initial_conditions}
\alias{make_initial_conditions}
\title{Make the initial conditions}
\usage{
make_initial_conditions(
  n_activity,
  initial_condition_rule,
  fraction_exposed,
  infected_class = NULL
)
}
\arguments{
\item{initial_condition_rule}{the rule used to}

\item{infected_class}{if the initial condition rule}

\item{frac_infected}{fraction infected at first}
}
\description{
uses the user-supplied initial condition rule
and the fraction infected, to generate initial
conditions for forward simulation
}
