% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_onto_noswitch.R
\name{make_initial_guess}
\alias{make_initial_guess}
\title{Make an initial guess of an 'effective' mixing matrix}
\usage{
make_initial_guess(mix_mat, switch_mat, parameterisation)
}
\arguments{
\item{mix_mat}{The contacts between different activity classes (and within)
those activity classes.}

\item{switch_mat}{Describes the transitions between the different activity
classes.}

\item{parameterisation}{The parameterisation (other than the mix/switch)
matrices. So, recovery rate, transmission rate, etc.}
}
\description{
Function takes the mixing and switching matrices to be
reproduced and applies an operator to them to make
the 'effective' mixing matrix. That is, the matrix that if
}
