% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vary_parameters.R
\name{make_initial_specific}
\alias{make_initial_specific}
\title{Specific activity class infected first.}
\usage{
make_initial_specific(n_activity, initial_class, fraction_exposed)
}
\arguments{
\item{n_activity}{The number of activity classes in the model.}

\item{initial_class}{the initial class which contains the exposeds}

\item{fraction_exposed}{The fraction of the activity class which
is moved to the Exposed
(E) compartment.}
}
\description{
initial condition where we only expose one activity class (user-specified)
rather than exposing all activity classes initially.
This has nontrivial consequences
for the ensuing dynamics,
}
