% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vary_parameters.R
\name{make_initial_uniform}
\alias{make_initial_uniform}
\title{Initial conditions where each compartment is inoculated simultaenously with
an equal fraction of exposed individuals (or, rather, stuff...)}
\usage{
make_initial_uniform(n_activity, fraction_exposed)
}
\arguments{
\item{n_activity}{the number of activity classes in the model}

\item{fraction_exposed}{the fraction of each activity class which is
exposed to the infection.}
}
\value{
list, whose items are \code{E0} and \code{S0}.
These arrays are of length (n_activity).
}
\description{
This is a simple initial condition,
Everything else is in the Susceptible
compartment.
}
