% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_curves.R
\name{make_metric_comparison}
\alias{make_metric_comparison}
\title{Metric comparison across models}
\usage{
make_metric_comparison(parsets)
}
\arguments{
\item{parsets}{A list of parameter sets.}
}
\description{
Computes the comparison metrics for a list of parameter sets.
}
