% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_curves.R
\name{make_parameter_sets}
\alias{make_parameter_sets}
\title{Make parameter sets for model fitting}
\usage{
make_parameter_sets(
  N,
  r0,
  num_activity,
  haslemere_activity_scores,
  path_to_empirical_mix = "data/Haslemere.csv"
)
}
\description{
Make parameter sets for model fitting
}
