% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_store.R
\name{make_parms}
\alias{make_parms}
\title{Make a default parameterisaiton}
\usage{
make_parms(
  n_activity,
  epsilon,
  switch_rate,
  initial_condition_rule,
  infected_class = NULL,
  infected_fraction,
  activity_scheme = "gamma",
  gamma_dist = NULL,
  mixing_model,
  switching_model,
  target_r0,
  power_param = NULL,
  pathogen_string = "sars-cov-2"
)
}
\arguments{
\item{n_activity}{number of activity classes}

\item{epsilon}{assortativity parameter for mixing matrix}

\item{switch_rate}{rate of switching between activity classes}

\item{initial_condition_rule}{rule for initial conditions. allowed values are
\code{uniformly_infected} and \code{specific_class_infected}}

\item{infected_class}{if \code{initial_condition_rule} is
\code{specific_class_infected},
this parameter specifies which class is initially infected.}

\item{activity_scheme}{scheme for activity scores. Allowed values are
\code{linear}, \code{null} and \code{gamma}. Default is \code{gamma}.
This specifies how activity scores
are generated for each activity class.}

\item{gamma_dist}{if \code{activity_scheme} is \code{gamma}, this parameter
is a list containing the parameters for the gamma distribution.
It can contain either
\code{mean} and \code{variance}, or \code{shape} and \code{rate}. (But not a mixture! :p)}

\item{power_param}{optional power parameter for the \code{polynomial} mixing model.}

\item{pathogen_string}{string describing the pathogen to be modelled.
Currently only \code{sars-cov-2} is implemented.}
}
\value{
a list of parameters for the simulation.
}
\description{
This function makes the parameterisation for mix-switch
simulation.
N (total population) is set to 1 million people by default.
Switching rate has default 0.1.
Initial conditions are set according to the specified rule
Function assumes that the class sizes are of equal size.
}
