% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_store.R
\name{make_switch_matrix}
\alias{make_switch_matrix}
\title{Make the switching matrix according to supplied model}
\usage{
make_switch_matrix(n_activity, switch_rate, switching_model)
}
\arguments{
\item{n_activity}{the number of activity classes in the odin model.}

\item{switch_rate}{the rate of switching. this = 1/Expected residence time
in a particular activity class}

\item{switching_model}{allowed values are \code{simple_switching} and \verb{adjacent_ switching}.}
}
\description{
Wrapper - mostly for concision - which sanitises the
switch rate input into the function and uses the supplied
model to produce the CTMC generator matrix supplied to odin
}
\examples{
# Example call
swch <- make_switch_matrix(
  n_activity = 5,
  switch_rate = 0.1,
  switching_model = "simple_switching"
)
print(swch)
# Example (adjacent wrap-around switching)
swch <- make_switch_matrix(
  n_activity = 5,
  switch_rate = 0.1,
  switching_model = "adjacent_switching"
)
print(swch)
}
\concept{switching_model}
