% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_onto_noswitch.R
\name{mat_from_vec}
\alias{mat_from_vec}
\title{Make a matrix suitable for use as a contact matrix from a supplied vector}
\usage{
mat_from_vec(vec_opt, ndim)
}
\arguments{
\item{vec_opt}{the vector used by the optimiser}

\item{ndim}{The number of activity classes used}
}
\description{
Make a matrix suitable for use as a contact matrix from a supplied vector
}
