% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piplelines.R
\name{metaroutine}
\alias{metaroutine}
\title{Run a standard set of routines}
\usage{
metaroutine(
  n_activity = 5,
  resol = 10,
  gamma_prmt = list(mean = 30, variance = 2500)
)
}
\arguments{
\item{n_activity}{the number of activity classes in the model.}

\item{resol}{the number of values in each direction to scan.}

\item{gamma_prmt}{the gamma distribution according to which the activity
scores are calculated.}
}
\description{
Wrapper for testing combinations of parameters / conditions.
}
