% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixSwitchEpi-package.R
\docType{package}
\name{MixSwitchEpi-package}
\alias{MixSwitchEpi}
\alias{MixSwitchEpi-package}
\title{MixSwitchEpi: Epidemics in Heterogeneously Active Populations With Movement Between Activity Classes}
\description{
Models and analysis of SEIR epidemics in heterogeneous populations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://vargasrichards.github.io/MixSwitchEpi/}
  \item \url{https://github.com/vargasrichards/MixSwitchEpi}
}

}
\author{
\strong{Maintainer}: Alexis Vargas Richards \email{amr625@ic.ac.uk}

}
\keyword{internal}
