% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_plots.R
\name{multi_outcome_comparison}
\alias{multi_outcome_comparison}
\title{Multi-outcome 2-D comparison heatmaps}
\usage{
multi_outcome_comparison(
  base_params,
  file,
  response_vars,
  ref_switch = NULL,
  ref_r0,
  n_activity = 5,
  output_dir = "output/comparison",
  save = TRUE,
  width = 6,
  height = 4,
  show_diff = TRUE,
  slice_res_times = NULL
)
}
\arguments{
\item{base_params}{list from \code{make_parms()}}

\item{file}{scan output}

\item{response_vars}{character vector of metric column names}

\item{ref_r0}{optional R0}

\item{n_activity}{the (integer) number of activity classes in the model}

\item{output_dir}{save location}

\item{save}{write files?}

\item{width, height}{size per panel}

\item{ref_epsilon, ref_switch}{reference point}
}
\description{
Multi-outcome 2-D comparison heatmaps
}
