% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_metrics.R
\name{peak_synchronicity}
\alias{peak_synchronicity}
\title{Compute how synchronised epidemic peaks are}
\usage{
peak_synchronicity(peak_data)
}
\arguments{
\item{peak_data}{Data frame of peak times per class}
}
\description{
Compute how synchronised epidemic peaks are
}
