% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_curves.R
\name{pivot_seir}
\alias{pivot_seir}
\title{Pivot SEIR epidemic df + prepare for plotting}
\usage{
pivot_seir(df, model_name)
}
\arguments{
\item{df}{the epidemic dataframe}

\item{model_name}{string descriptor for the model}
}
\description{
Pivots the epidemic dataframe into suitable plotting format
and then binds an epidemic label
}
