% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{plot_2d_heatmaps}
\alias{plot_2d_heatmaps}
\title{Plot 2D heatmaps of epidemic metrics}
\usage{
plot_2d_heatmaps(scan_results, metrics_to_plot = NULL, overall_only = TRUE)
}
\arguments{
\item{scan_results}{Results from scan_2d_parameter_space()}

\item{metrics_to_plot}{Vector of metric names to plot (default: all)}

\item{overall_only}{If TRUE, only plot overall metrics (class == -1)}
}
\value{
List of ggplot objects
}
\description{
Plot 2D heatmaps of epidemic metrics
}
