% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piplelines.R
\name{plot_combin}
\alias{plot_combin}
\title{Plot combined results}
\usage{
plot_combin(results_file)
}
\arguments{
\item{results_file}{the results file from
which the results to be plotted are read.}
}
\description{
Plot the results from a saved CSV file.
}
\details{
Makes a heatmap by residency time.
}
