% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{plot_epidemic_with_rt}
\alias{plot_epidemic_with_rt}
\title{Combined plot of epidemic trajectory and Rt}
\usage{
plot_epidemic_with_rt(
  model_output,
  computed_metrics,
  compartments = c("I", "E"),
  title = "Epidemic Dynamics and Rt",
  rt_only = FALSE,
  axis_text_size,
  axis_title_size
)
}
\arguments{
\item{model_output}{Model output with Rt column (from compute_rt_timeseries)}

\item{computed_metrics}{Computed epidemic metrics - this allows for}

\item{title}{the title of the plot}

\item{rt_only}{whether to plot only the evolution of Rt with time}
}
\description{
Creates a two-panel plot showing the epidemic course (I, E) and Rt
}
