% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_analysis.R
\name{plot_metric_mds}
\alias{plot_metric_mds}
\title{Plot MDS of metrics}
\usage{
plot_metric_mds(analysis_res)
}
\arguments{
\item{analysis_res}{Result from analyse_metric_dimensionality}
}
\description{
Plots the multidimensional scaling projection of the metrics themselves
to observe clustering and find the minimal set of descriptors.
}
