% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_analysis.R
\name{plot_metric_pairs}
\alias{plot_metric_pairs}
\title{Plot Pairwise Scatterplots of Metrics}
\usage{
plot_metric_pairs(analysis_res, color_var = "epsilon")
}
\arguments{
\item{analysis_res}{Result from analyse_metric_dimensionality}

\item{color_var}{A column name from the original data to color by (e.g. 'epsilon')}
}
\description{
Plot Pairwise Scatterplots of Metrics
}
