% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_analysis.R
\name{plot_metric_pca}
\alias{plot_metric_pca}
\title{Plot PCA biplot of metrics}
\usage{
plot_metric_pca(analysis_res, color_var = "epsilon")
}
\arguments{
\item{analysis_res}{Result from analyse_metric_dimensionality}

\item{color_var}{Variable to color the spatial points (e.g. 'epsilon', 'switch_rate')}
}
\description{
Plot PCA biplot of metrics
}
