% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{plot_sweep_results_1d}
\alias{plot_sweep_results_1d}
\title{Plot 1D parameter sweeps}
\usage{
plot_sweep_results_1d(results_tibble)
}
\arguments{
\item{results_tibble}{the results tibble containing parameter scan results}
}
\description{
Plot 1D parameter sweeps
}
