% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_models.R
\name{polynomial_mixing}
\alias{polynomial_mixing}
\title{Parametrically generate mixing matrix for the desired}
\usage{
polynomial_mixing(
  n_activity,
  power_param,
  epsilon_assort,
  act_vector,
  class_size_vector
)
}
\arguments{
\item{n_activity}{the number of activity classes in the model}

\item{power_param}{the power parameter. if = 1}

\item{epsilon_assort}{the assortativity.}

\item{act_vector}{the vector of activity class scores.
Gives a measure of the activity of each class.}

\item{class_size_vector}{the vector of activity class sizes}
}
\description{
Note that this has an extra parameter compared to the other mixing
models. Thus the call signature needs to be adjusted accordingly
}
