% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vary_parameters.R
\name{report_scan_results}
\alias{report_scan_results}
\title{Report the parameter scan results}
\usage{
report_scan_results(results_tibble, parameter_tibble, create_dir = FALSE)
}
\arguments{
\item{results_tibble}{a tibble/df containing the results of a parameter scan}

\item{parameter_tibble}{the tibble containing the parameters used for that
scan.}
}
\description{
Writes out the summary statistics for the parameter scan,
along with the time at which
started and completed. Makes a directory for this.
}
\details{
perhaps should be updated to measure the amount of cpu time the scan takes?
}
