% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_metrics.R
\name{rho_entry}
\alias{rho_entry}
\title{Generates a single entry of the rho matrix}
\usage{
rho_entry(i, j, params)
}
\arguments{
\item{i}{the row index}

\item{j}{the column index}

\item{params}{the model parameters}
}
\description{
Generates a single entry of the rho matrix
}
