% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_metrics.R
\name{rho_matrix}
\alias{rho_matrix}
\title{Generates the rho matrix from parameters}
\usage{
rho_matrix(params)
}
\arguments{
\item{params}{the model parameters}
}
\description{
The matrix of mixing probabilities between activity classes. This is the matrix of
probabilities that an individual in activity class i has contact with an
individual in activity class j.
See the documentation for \code{rho_entry} for details on how this is constructed.
}
