% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piplelines.R
\name{run_combination}
\alias{run_combination}
\title{Run and record a 2D parameter scan}
\usage{
run_combination(
  n_activity,
  switch_model,
  mix_model,
  resol,
  gamma_prmt,
  initial,
  calib_eps,
  calib_swch,
  target_r0
)
}
\arguments{
\item{n_activity}{the number of activity classes in the model.}

\item{switch_model}{switching model}

\item{mix_model}{mixing model}

\item{resol}{the number of values in each direction to scan.
so the total number of parameter combinations run will be resol^2.}

\item{gamma_prmt}{the gamma distribution according to which the activity
scores are calculated.}

\item{initial}{initial condition rule. allowed values include "uniform",}

\item{calib_eps}{the epsilon at which to calibrate to target_r0}

\item{calib_swch}{the switch rate at which to calibrate t0 target_r0}

\item{target_r0}{the R0 to calibrate to.}
}
\description{
A wrapper for parameter set construction, calibration,
simulation and writing out results.
}
