% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir_mixswitch.R
\name{run_system_wparms}
\alias{run_system_wparms}
\title{Run the system and produce neat output}
\usage{
run_system_wparms(parms, model)
}
\arguments{
\item{parms}{parameters for the model}

\item{model}{odin model}
}
\description{
Bare-bones function to simulate the system using dust.
can call this many times if needed, for eg optimisation
}
\concept{simulation}
