% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{save_heatmaps}
\alias{save_heatmaps}
\title{Save all heatmap plots}
\usage{
save_heatmaps(plot_list, output_dir = "heatmaps", width = 8, height = 6)
}
\arguments{
\item{plot_list}{List of ggplot objects from plot_2d_heatmaps()}

\item{output_dir}{Directory to save plots}

\item{width}{Plot width in inches}

\item{height}{Plot height in inches}
}
\description{
Save all heatmap plots
}
