% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_metrics.R
\name{scale_results}
\alias{scale_results}
\title{Scale results by occupancy fraction for each activity class}
\usage{
scale_results(model_output)
}
\arguments{
\item{model_output}{Output from dust/odin run}
}
\description{
Scale results by occupancy fraction for each activity class
}
\seealso{
Other epidemic_metrics: 
\code{\link{characterise_end}()},
\code{\link{characterise_peaks}()},
\code{\link{compute_all}()},
\code{\link{compute_hit}()},
\code{\link{compute_rt}()},
\code{\link{compute_spectral_gap}()},
\code{\link{compute_ttprev}()},
\code{\link{find_peaks}()}
}
\concept{epidemic_metrics}
