% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vary_parameters.R
\name{scan_mixswitch_framework}
\alias{scan_mixswitch_framework}
\title{Scan 2D parameter space (epsilon × switch_rate.}
\usage{
scan_mixswitch_framework(
  base_pars,
  epsilon_vals,
  switch_vals,
  n_epsilon = 11,
  n_switch = 11,
  classed_detail = FALSE,
  switch_model = "simple_switching",
  mix_model,
  power_param = NULL,
  initial_condition_rule = NULL,
  frac_infected_initial = 1e-04,
  calibration_mode = "none",
  t_end = 1200
)
}
\arguments{
\item{base_pars}{Calibrated base parameters from default_setup().}

\item{epsilon_vals}{Vector of epsilon values to scan (default: 0 to 1)}

\item{switch_vals}{Vector of switching rate values to scan (default: 0 to 1)}

\item{n_epsilon}{Number of epsilon points (if epsilon_vals not provided)}

\item{n_switch}{Number of switch rate points (if switch_vals not provided)}

\item{classed_detail}{Whether to compute class-specific metrics}

\item{switch_model}{the model of switching to be used.
allowed values include:}

\item{mix_model}{the model for}

\item{initial_condition_rule}{the rule  to generate the initial cond.}

\item{frac_infected_initial}{the initial fraction infected of the population
This will be put in the E (exposed) compartment.}

\item{calibration_mode}{mode of calibration. allowed values are 'none', 'switch', 'assort', 'both'.
If 'none', no recalibration. If 'switch', calibrate once for each switch value to the target r0.
If 'assort', dont calibrate for switch but calibrate for assort. If 'both', calibrate each point in mix-switch space.}

\item{t_end}{simulation horizon in days (default 1200)}
}
\value{
data.table with all epidemic metrics for each parameter combination
}
\description{
Basic flexible framework function which can be adapted significantly
as we require. Runs a set of workers in parallel.
We precompute the mixing and switching matrices to reduce
computational time.
}
