% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_onto_noswitch.R
\name{score_difference}
\alias{score_difference}
\title{Score difference between model outputs}
\usage{
score_difference(x, reference_pattern, shared_parms, model, similar_matrices)
}
\arguments{
\item{x}{the vector used by the optimiser to represent the mixing matrix.}

\item{reference_pattern}{gives the desired pattern of e.g., E(t) through
time, which we're going to try and match as close as poss. with no switching.}

\item{shared_parms}{the shared parameters (e.g., transmission rate, recovery
rate).}

\item{model}{the odin/monty model}

\item{similar_matrices}{bool. Indicates whether to
reward effective contact matrices
which are similar to the original contact matrix. Experimental feature.}
}
\description{
Produces a score of the difference in model result
between the original mixing/switching parameterisation (\code{true_matrices})
and the candidate mixing matrix.
}
