% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_metrics.R
\name{sigma_matrix}
\alias{sigma_matrix}
\title{Makes the sigma matrix}
\usage{
sigma_matrix(params)
}
\arguments{
\item{params}{the model parameters}
}
\description{
Constructs the matrix of the state-transitions.
}
\seealso{
Other r0_calc: 
\code{\link{compute_ngm}()},
\code{\link{compute_r0}()}
}
\concept{r0_calc}
