% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_models.R
\name{simple_switching}
\alias{simple_switching}
\title{Make the simplest possible switching matrix.}
\usage{
simple_switching(switch_rate, num_classes)
}
\arguments{
\item{switch_rate}{rate of switching between activity classes}

\item{num_classes}{number of activity classes}
}
\value{
A square matrix of dimension num_classes x num_classes
}
\description{
User specifies the number of activity classes to be used and the
switching rate = 1 - rate(remain) in class i, conditional on not
recovering.
}
