% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_metrics.R
\name{spectral_radius}
\alias{spectral_radius}
\title{Convienience function to compute the spectral radius of a matrix}
\usage{
spectral_radius(a)
}
\arguments{
\item{a}{matrix}
}
\value{
the spectral radius (max abs eigenvalue)
}
\description{
The spectral radius is the maximum absolute value of the eigenvalues of the matrix.
}
\seealso{
\code{compute_r0}, \code{compute_rt}
}
