% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_metrics.R
\name{t_entry}
\alias{t_entry}
\title{Generates a single entry of the T matrix}
\usage{
t_entry(i, j, params)
}
\description{
The rate of transmission from class j to class i
}
\seealso{
\code{t_matrix}
}
