% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_metrics.R
\name{t_matrix}
\alias{t_matrix}
\title{Generates the T matrix from parameters}
\usage{
t_matrix(params)
}
\arguments{
\item{params}{the model parameters}
}
\description{
This T matrix describes the transmission component of the
NGM. See Mathematical Tools for Understanding Infectious Disease Dynamics,
Diekmann et al  for details.
}
\seealso{
\code{t_entry}
}
