% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_onto_noswitch.R
\name{vec_from_mat}
\alias{vec_from_mat}
\title{Make a vector for input into optimisation fun.}
\usage{
vec_from_mat(M)
}
\arguments{
\item{M}{the matrix from which we get the x for optimisation}
}
\description{
Make a vector for input into optimisation fun.
}
